Moved mem from src2 -> src
